sync_wait() remove extra move constructor call (#302)
* sync_wait() remove extra move constructor call

User @baderouaich identified that libcoro's sync_wait on a complex
object would invoke an extra move constructor that other coroutine
libraries did not. This change now correctly forwards the return_value
into the promise if the object is move constructible instead of double
moving.

Closes #286

* fix windows build issues

+++ test/test_thread_pool.cpp
@@ -183,7 +183,7 @@ TEST_CASE("thread_pool high cpu usage when threadcount is greater than the numbe
 
     auto wait_for_task = [](coro::thread_pool& pool, std::chrono::seconds delay) -> coro::task<>
     {
-        auto sleep_for_task = [](std::chrono::seconds duration) -> coro::task<int>
+        auto sleep_for_task = [](std::chrono::seconds duration) -> coro::task<int64_t>
         {
             std::this_thread::sleep_for(duration);
             co_return duration.count();
